[염정운] PageTop.vue를 두 개의 컴포넌트로 쪼개고 PageTop.vue는 삭제
